Refactor metrics processing
Setting metrics was tightly coupled with the gathering of the data.
This made it difficult to change the logic without a significant amount
of code changes.

By splitting the setting of the metrics into its own function, we can
change the high level logic without having to move hundreds of lines
around.

Co-authored-by: Lukas Eklund <leklund@gmail.com>
Signed-off-by: Michael Lorant <michael.lorant@nine.com.au>